Agregue la funcion de healing

+++ src/main/java/movimientos/Healing.java
@@ -2,16 +2,15 @@ package movimientos;
 
 public class Healing extends Movimientos{
 
-    private float heal;
-    public Healing(String nombre, int accuracy, int PP,float heal, String tipo) {
+    private int heal;
+    public Healing(String nombre, int accuracy, int PP,int heal, String tipo) {
         super(nombre, accuracy, PP, tipo);
         this.heal=heal;
     }
-    public float getHeal() {
+    public int getHeal() {
         return heal;
     }
     public void setHeal(int heal) {
         this.heal = heal;
     }
-
 }
